src/main/java/br/edu/ifce/tjw/serviceImpl/RoleServiceImpl.java
package br.edu.ifce.tjw.serviceImpl;

import br.edu.ifce.tjw.dao.RoleDao;
import br.edu.ifce.tjw.model.Role;
import br.edu.ifce.tjw.service.RoleService;

import org.springframework.stereotype.Service;
import org.springframework.transaction.annotation.Transactional;

@Service
@Transactional
public class RoleServiceImpl implements RoleService {

    private RoleDao roleDao;

    public RoleServiceImpl(RoleDao roleDao) {
        this.roleDao = roleDao;
    }

    @Override
    public Role loadRoleByName(String roleName) {
        return roleDao.findByName(roleName);
    }

    @Override
    public Role createRole(String roleName) {
        return roleDao.save(new Role(roleName));
    }
}
